requirements.txt

asgiref==3.4.1
astroid==2.9.3
backports.zoneinfo==0.2.1
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.10
defusedxml==0.7.1
Django==4.0.1
django-filter==21.1
django-rest-auth==0.9.5
djangorestframework==3.13.1
idna==3.3
importlib-metadata==4.10.0
isort==5.10.1
lazy-object-proxy==1.7.1
Markdown==3.3.6
mccabe==0.6.1
oauthlib==3.1.1
platformdirs==2.4.1
psycopg2==2.9.3
pylint==2.12.2
pylint-django==2.5.0
pylint-plugin-utils==0.7
python3-openid==3.2.0
pytz==2021.3
requests==2.27.1
requests-oauthlib==1.3.0
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
typing-extensions==4.0.1
urllib3==1.26.8
wrapt==1.13.3
zipp==3.7.0